% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{asHCVdata}
\alias{asHCVdata}
\title{asHCVdata function}
\usage{
asHCVdata(dat, hcv, subjectid, gridnum, x, y)
}
\arguments{
\item{dat}{a dataframe with the HCV data, sorted in 'expand.grid' order within patient and sample number}

\item{hcv}{a character string, the name of the variable containing the HCV levels}

\item{subjectid}{a character string, the name of the variable containing the subject identifier}

\item{gridnum}{a character string, the name of the variable containing the sample number for each individual}

\item{x}{a character string, the name of the variable containing the x-index of the grid}

\item{y}{a character string, the name of the variable containing the y-index of the grid}
}
\value{
...
}
\description{
A function to convert a dataset into one that can be used with HCV's internal/visible functions
}
