% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{bootCI}
\alias{bootCI}
\title{bootCI function}
\usage{
bootCI(param, nboot, dat, iForm = NULL, sForm = NULL, nr = NULL,
  nc = NULL)
}
\arguments{
\item{param}{backtransformed parameters}

\item{nboot}{number of bootstrap samples}

\item{dat}{a HCVdataset object, see ?asHCVdata}

\item{iForm}{formula for infectivity}

\item{sForm}{formula for susceptibility}

\item{nr}{integer number of rows in grid}

\item{nc}{integer number of columns in grid}
}
\value{
matrix containing bootstrapped parameter estimates
}
\description{
A function to compute boostrapped confidence intervals for the partial likelihood model
}
