\name{d4}
\docType{data}
\alias{d4}
\title{Simulated HCV Data 2018-04-11}
\description{Simulated HCV Data 2018-04-11}
\usage{d4}
\format{matrix}
\source{LANL}
\references{Graw et al. (2014) Inferring Viral Dynamics in Chronically HCV Infected Patients from the Spatial Distribution of Infected Hepatocytes}
\keyword{datasets}
