% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{distFun}
\alias{distFun}
\title{distFun function}
\usage{
distFun(mat)
}
\arguments{
\item{mat}{matrix of hcv levels vectorised using function as.vector, see prepData function}
}
\value{
a function to evaluate the transmission probabilities from cell i to cell j
}
\description{
An internal function to create a function to evaluate the transmission probabilities from cell i to cell j - includes model for susceptibility and also infectivity, as well as distance decay
}
