% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{gammaFun}
\alias{gammaFun}
\title{gammaFun function}
\usage{
gammaFun(dst, theta, phi)
}
\arguments{
\item{dst}{distance between cells}

\item{theta}{paramter theta}

\item{phi}{parameter phi}
}
\value{
the function gamma evaluated at those distances
}
\description{
An internal two-parameter function to evaluate the measure of dependence between cells a given distance apart. Function must work on vector inputs.
}
\details{
On the to-do list is extend this to more general functions with an arbitrary number of parameters
}
