% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{getSI}
\alias{getSI}
\title{getSI function}
\usage{
getSI(mat)
}
\arguments{
\item{mat}{matrix of infection times: assumes there are no ties in these times.}
}
\value{
a list of infected/susceptibles at each time point
}
\description{
An internal function to compute susceptible/infected given a matrix of infection times: assumes there are no ties in these times.
}
