% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCV.R
\docType{package}
\name{HCV-package}
\alias{HCV-package}
\alias{spatsurv}
\title{HCV}
\format{An object of class \code{logical} of length 1.}
\usage{
spatsurv
}
\description{
An R package for
}
\details{
\tabular{ll}{
Package: \tab spatsurv\cr
Version: \tab 0.9-11\cr
Date: \tab 2015-06-24\cr
License: \tab GPL-3 \cr
}



section{Dependencies}{
To be filled in at a later date
}

section{Citation}{
No citation Info yet.
}

references{
X
}
}
\author{
Paula E. Moraga-Serrano, CHICAS, Department of Medicine, Lancaster University,
 Benjamin M. Taylor, CHICAS, Department of Medicine, Lancaster University
}
\keyword{package}
