% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{logPL}
\alias{logPL}
\title{logPL function}
\usage{
logPL(param, si, distfun, inf, sus, N)
}
\arguments{
\item{param}{vector of parameters on the transformed scale (note they are back-transformed as part of this routine)}

\item{si}{list of susceptible/infected, see ?getSI}

\item{distfun}{function created using the function distFun}

\item{inf}{covariate data relating to infectivity}

\item{sus}{covariate data relating to susceptibility}

\item{N}{vector of length 2: the number of covariates associated with respectively infectivity and susceptibility}
}
\value{
the log-partial likelihood evaluated at the input parameters
}
\description{
An internal function to calculate the partial likelihood
}
