% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{optimPL}
\alias{optimPL}
\title{optimPL function}
\usage{
optimPL(dat, iForm = NULL, sForm = NULL, inits = NULL, pd = NULL,
  ...)
}
\arguments{
\item{dat}{a HCVdataset object, see ?asHCVdata}

\item{iForm}{formula for infectivity}

\item{sForm}{formula for susceptibility}

\item{inits}{initial values to pass to optim}

\item{pd}{optional prepData output, used in function bootCI}

\item{...}{addition arguments passed to optim function}
}
\value{
output from optim, run on the (log) partial likelihood function
}
\description{
A function to maximise the partial likelihood.
}
