% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{partrans}
\alias{partrans}
\title{partrans function}
\usage{
partrans(param)
}
\arguments{
\item{param}{a vector of transformed parameters (theta,phi,beta_1,...,beta_d)}
}
\value{
a vector of back-transformed parameters
}
\description{
An internal function to backtransform the parameters
}
