% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{prepData}
\alias{prepData}
\title{prepData function}
\usage{
prepData(dat, iForm = NULL, sForm = NULL, inits = NULL)
}
\arguments{
\item{dat}{a HCVdataset object, see ?asHCVdata}

\item{iForm}{formula for infectivity}

\item{sForm}{formula for susceptibility}

\item{inits}{initial values for optimiser, the default is NULL, where initial values are chosen automatically (in an ad-hoc fashion)}
}
\value{
the data and model inputs ready to run the optimiser.
}
\description{
An internal function to prepare data into a form ready for analysis. Many computations with this model can be performed once at the start and stored in memory in order to increase computational speed; this function manages that process.
}
