% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_functions_with_covariates.R
\name{simulateProcess}
\alias{simulateProcess}
\title{simulateProcess function}
\usage{
simulateProcess(param, startcell, nmax, nrow = 10, ncol = 10,
  iForm = NULL, sForm = NULL, dat = NULL)
}
\arguments{
\item{param}{vector of back transformed parameters}

\item{startcell}{index of cell with first infection}

\item{nmax}{number of cells to be infected}

\item{nrow}{number of rows in grid}

\item{ncol}{number of columns in grid}

\item{iForm}{formula for infectivity}

\item{sForm}{formula for susceptibility}

\item{dat}{a HCVdataset object, see ?asHCVdata, this will contain the covariate data for simulation}
}
\value{
an output grid of HCV values: larger values represent older infections
}
\description{
A function to simulate a realisation of a process, only works with one individual grid of data ATM
}
