% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCV.R
\name{SMI}
\alias{SMI}
\title{Spatial Mixture Index}
\usage{
HierarchicalVoronoi(constraint_domain, optimization_domain, hclsut, max_k)
}
\arguments{
\item{constraint_domain:}{`[matrix]`: The geometry location of the data.}

\item{optimization_domain:}{`[matrix]`: The spatial feature of the data.}

\item{hclsut:}{`[hclust]`: A hclsut object generate by function `hclust` or `HierarchicalVoronoi`}
}

\item{max_k:}{`[integer]`: A number which indicate the maximum number of group}
}

}
\value{
A 'list' object. The attribute of SMI is a list with component:

bestClsuter : The optimal number of clusters from 2 to `max_k`.

index : The SMI index values of clusters from 2 to `max_k`.

ratio : The different ratio for SMI index value.
}
\description{
Determining the optiaml number of clusters under the constraint of spatial homogeneity.
}
\examples{
# data <- synthetic_data(5, 4, 0.15, 1000, 2)
# result <- HierarchicalVoronoi(data$geo, data$feat, 'ward.D', 2)
# SMIres <- SMI(data$geo, data$feat, result, 10)
}
\author{
DongDong-Zoez <lbry5230100@gmail.com> University of Taiwan NSYSU.
}
\seealso{
`synthetic_data`, `HierarchicalVoronoi`
}
\details{
The SMI function is an internal index for involving the spatial homogeneity. To measure the compactness of the domain, we assume that the compactness is measured by the linear combination of within-group sum of squared of constraint domain and optimization domain, the linear coefficient is determined by the average length of Delaunay edges for each group. Based on the idea, SMI used different average length mean to estimate the degeree of spatial homogeneity, then implement the Sigmoid function for smoothing the result.
}

