% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCV.R
\name{HierarchicalVoronoi}
\alias{HierarchicalVoronoi}
\title{The hierarchical clustering based on Voronoi common edges}
\usage{
HierarchicalVoronoi(constraint_domain, optimization_domain, linkage, iterate)
}
\arguments{
\item{constraint_domain:}{`[matrix]`: The geometry location of the data.}

\item{optimization_domain:}{`[matrix]`: The spatial feature of the data.}

\item{linkage:}{`[character]`: The valid arguments are "single", "complete", "average", "median", "ward", "centroid" and "weighted".}

\item{iterate:}{`[integer]`: The number of clusters, iterate = 2 is recommanded.}
}
}
\value{
a 'hclust' object.
}
\description{
The hierarchical clustering methods for the constraint of spatial homogeneity.
}
\examples{
# data <- synthetic_data(5, 4, 0.15, 1000, 2)
# clustStructure <- HierarchicalVoronoi(data$geo, data$feat, 'ward.D', 2)
# clustStructure$label_matrix[1,] # The cluster label with number of clusters equal to `iterate`.
}
\author{
DongDong-Zoez <lbry5230100@gmail.com> University of Taiwan NSYSU.
}
\seealso{
`synthetic_data`
}
\details{
The function is used to solve the dual problem. A dual problem is to minimize the cost of the spatial feature domain or the domain of your data attribute, at the meanwhile, we are expect that the cluster of geometry domain or constraint domain will form a compact region, namely, each cluster is bounded by some irregular curves and not overlap for each pairs of clusters. To achieve the goal, one is to consider the spatial homogeneity for spatial data, and this paper is following the idea of constructing the Delaunay triangulation and Voronoi diagram to make an approach for spatially homogeneity.
}

